requirements.txt
certifi==2019.3.9
chardet==3.0.4
dropbox==9.3.0
idna==2.8
mysql-connector==2.2.9
prettytable==0.7.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
